Export cookies conversion method
