+++ packages/record/src/index.ts
@@ -1,3 +1,4 @@
 export { recordSession } from "./record/record";
 export { recordLoginFlowSession } from "./record/record-login-flow";
 export { bootstrapPage as bootstrapRecordingPage } from "./record/record.utils";
+export { convertCookieToMeticulous } from "./record/utils/cookie-helpers";
