+++ packages/record/src/record/utils/provide-cookie-access.ts
@@ -1,5 +1,6 @@
 import { Cookie } from "@alwaysmeticulous/api";
-import { Page, Protocol } from "puppeteer";
+import { Page } from "puppeteer";
+import { convertCookieToMeticulous } from "./cookie-helpers";
 
 type ModifiedWindow = {
   __meticulous?: {
@@ -12,14 +13,7 @@ export const provideCookieAccess = async (page: Page) => {
   const client = await page.target().createCDPSession();
   const getCookies = async () =>
     (await client.send("Network.getAllCookies")).cookies.map(
-      ({ sameSite, expires, ...rest }): Cookie => {
-        const convertedSameSite = convertSameSiteValue(sameSite);
-        return {
-          ...rest,
-          ...(convertedSameSite ? { sameSite: convertedSameSite } : {}),
-          expires: expires ? expires * 1000 : null, // Convert seconds to milliseconds
-        };
-      }
+      convertCookieToMeticulous
     );
   await page.exposeFunction("__meticulous_getAllCookies", getCookies);
 
@@ -32,19 +26,4 @@ export const provideCookieAccess = async (page: Page) => {
     ).__meticulous_getAllCookies!;
     (window as ModifiedWindow).__meticulous = meticulousObject;
   });
-};
-
-const convertSameSiteValue = (
-  sameSite: Protocol.Network.CookieSameSite | undefined
-) => {
-  switch (sameSite) {
-    case "None":
-      return "none";
-    case "Lax":
-      return "lax";
-    case "Strict":
-      return "strict";
-    default:
-      return undefined;
-  }
 };
